Move engine creation into intializeEngine.ts
A lot of modules depend on Engine.ts. Moving dependencies for creating and initializing the Engine out of Engine.ts makes it easier to avoid cyclical dependencies.
