+++ packages/client/src/route/public.tsx
@@ -5,7 +5,6 @@ import ErrorBoundary from '@xrengine/client-core/src/common/components/ErrorBoun
 import { LoadingCircle } from '@xrengine/client-core/src/components/LoadingCircle'
 import { ClientTransportHandler } from '@xrengine/client-core/src/transports/SocketWebRTCClientTransport'
 import { AuthService } from '@xrengine/client-core/src/user/services/AuthService'
-import { createEngine } from '@xrengine/engine/src/ecs/classes/Engine'
 import { initializeBrowser } from '@xrengine/engine/src/initializeEngine'
 import { Network } from '@xrengine/engine/src/networking/classes/Network'
