+++ packages/engine/src/avatar/AvatarInputSchema.test.ts
@@ -5,9 +5,10 @@ import { FollowCameraComponent, FollowCameraDefaultValues } from '../camera/comp
 import { TargetCameraRotationComponent } from '../camera/components/TargetCameraRotationComponent'
 import { LifecycleValue } from '../common/enums/LifecycleValue'
 import { NumericalType } from '../common/types/NumericalTypes'
-import { createEngine, Engine } from '../ecs/classes/Engine'
+import { Engine } from '../ecs/classes/Engine'
 import { addComponent, getComponent } from '../ecs/functions/ComponentFunctions'
 import { createEntity } from '../ecs/functions/EntityFunctions'
+import { createEngine } from '../initializeEngine'
 import { InputType } from '../input/enums/InputType'
 import { VectorSpringSimulator } from '../physics/classes/springs/VectorSpringSimulator'
 import { CollisionGroups } from '../physics/enums/CollisionGroups'
