+++ packages/engine/src/ecs/classes/Engine.ts
@@ -1,40 +1,14 @@
 import { AudioListener, Object3D, OrthographicCamera, PerspectiveCamera, Scene, XRFrame } from 'three'
 
-import { UserId } from '@xrengine/common/src/interfaces/UserId'
-import { addActionReceptor, createHyperStore, registerState } from '@xrengine/hyperflux'
+import type { UserId } from '@xrengine/common/src/interfaces/UserId'
+import { createHyperStore } from '@xrengine/hyperflux'
 
-import { InputValue } from '../../input/interfaces/InputValue'
-import { WorldState } from '../../networking/interfaces/WorldState'
-import { EngineRenderer } from '../../renderer/WebGLRendererSystem'
-import { ObjectLayers } from '../../scene/constants/ObjectLayers'
-import { createWorld, World } from '../classes/World'
-import { accessEngineState, EngineEventReceptor } from './EngineService'
-import { Entity } from './Entity'
-
-export const CreateEngine = Symbol('CreateEngine')
-
-/**
- * Creates a new instance of the engine. This initializes all properties and state for the engine,
- * adds action receptors and creates a new world.
- * @returns {Engine}
- */
-export function createEngine() {
-  return Engine[CreateEngine]()
-}
+import type { InputValue } from '../../input/interfaces/InputValue'
+import type { World } from '../classes/World'
+import type { Entity } from './Entity'
 
 export class Engine {
-  static [CreateEngine] = () => new Engine()
   static instance: Engine
-
-  constructor() {
-    Engine.instance = this
-    this.currentWorld = createWorld()
-    this.scene = new Scene()
-    this.scene.layers.set(ObjectLayers.Scene)
-    EngineRenderer.instance = new EngineRenderer()
-    registerState(this.currentWorld.store, WorldState)
-    addActionReceptor(this.store, EngineEventReceptor)
-  }
 
   /** The uuid of the logged-in user */
   userId: UserId
@@ -67,10 +41,6 @@ export class Engine {
    * Reference to the three.js scene object.
    */
   scene: Scene = null!
-  get sceneLoaded() {
-    return accessEngineState().sceneLoaded.value
-  }
-  sceneLoadPromises: Promise<void>[] = []
 
   /**
    * Map of object lists by layer
@@ -93,10 +63,6 @@ export class Engine {
 
   inputState = new Map<any, InputValue>()
   prevInputState = new Map<any, InputValue>()
-
-  get isInitialized() {
-    return accessEngineState().isEngineInitialized.value
-  }
 
   publicPath: string = null!
