+++ packages/engine/src/input/systems/ClientInputSystem.test.ts
@@ -4,9 +4,10 @@ import { AvatarInputSchema } from '../../avatar/AvatarInputSchema'
 import { BinaryValue } from '../../common/enums/BinaryValue'
 import { LifecycleValue } from '../../common/enums/LifecycleValue'
 import { NumericalType } from '../../common/types/NumericalTypes'
-import { createEngine, Engine } from '../../ecs/classes/Engine'
+import { Engine } from '../../ecs/classes/Engine'
 import { addComponent } from '../../ecs/functions/ComponentFunctions'
 import { createEntity } from '../../ecs/functions/EntityFunctions'
+import { createEngine } from '../../initializeEngine'
 import { InputComponent } from '../components/InputComponent'
 import { GamepadAxis } from '../enums/InputEnums'
 import { InputType } from '../enums/InputType'
