+++ packages/engine/src/interaction/systems/InteractiveSystem.test.ts
@@ -1,7 +1,7 @@
 import assert from 'assert'
 
 import { ParityValue } from '../../common/enums/ParityValue'
-import { createEngine, Engine } from '../../ecs/classes/Engine'
+import { Engine } from '../../ecs/classes/Engine'
 import { Entity } from '../../ecs/classes/Entity'
 import { addComponent, hasComponent } from '../../ecs/functions/ComponentFunctions'
 import { createEntity } from '../../ecs/functions/EntityFunctions'
