+++ packages/engine/src/scene/functions/loaders/AudioFunctions.test.ts
@@ -7,10 +7,11 @@ import { ComponentJson } from '@xrengine/common/src/interfaces/SceneInterface'
 import { AssetLoader } from '../../../assets/classes/AssetLoader'
 import { AudioComponent, AudioComponentType } from '../../../audio/components/AudioComponent'
 import { AudioType, AudioTypeType } from '../../../audio/constants/AudioConstants'
-import { createEngine, Engine } from '../../../ecs/classes/Engine'
+import { Engine } from '../../../ecs/classes/Engine'
 import { Entity } from '../../../ecs/classes/Entity'
 import { addComponent, getComponent, hasComponent } from '../../../ecs/functions/ComponentFunctions'
 import { createEntity } from '../../../ecs/functions/EntityFunctions'
+import { createEngine } from '../../../initializeEngine'
 import { EntityNodeComponent } from '../../components/EntityNodeComponent'
 import { ErrorComponent } from '../../components/ErrorComponent'
 import { MediaComponent, MediaComponentType } from '../../components/MediaComponent'
