+++ packages/engine/src/scene/functions/loaders/PointLightFunctions.test.ts
@@ -3,9 +3,9 @@ import { Color, PointLight } from 'three'
 
 import { ComponentJson } from '@xrengine/common/src/interfaces/SceneInterface'
 
-import { createEngine } from '../../../ecs/classes/Engine'
 import { getComponent, hasComponent } from '../../../ecs/functions/ComponentFunctions'
 import { createEntity } from '../../../ecs/functions/EntityFunctions'
+import { createEngine } from '../../../initializeEngine'
 import { Object3DComponent } from '../../components/Object3DComponent'
 import { deserializePointLight } from './PointLightFunctions'
