+++ packages/engine/src/scene/functions/loaders/RenderSettingsFunction.test.ts
@@ -2,8 +2,8 @@ import assert from 'assert'
 
 import { ComponentJson } from '@xrengine/common/src/interfaces/SceneInterface'
 
-import { createEngine } from '../../../ecs/classes/Engine'
 import { createEntity } from '../../../ecs/functions/EntityFunctions'
+import { createEngine } from '../../../initializeEngine'
 import { EngineRenderer } from '../../../renderer/WebGLRendererSystem'
 import { deserializeRenderSetting } from './RenderSettingsFunction'
