+++ packages/engine/tests/SoA.test.ts
@@ -2,9 +2,9 @@ import assert from 'assert'
 import { Vector3 } from 'three'
 
 import { createQuaternionProxy, createVector3Proxy } from '../src/common/proxies/three'
-import { createEngine } from '../src/ecs/classes/Engine'
 import { addComponent } from '../src/ecs/functions/ComponentFunctions'
 import { createEntity } from '../src/ecs/functions/EntityFunctions'
+import { createEngine } from '../src/initializeEngine'
 import { TransformComponent } from '../src/transform/components/TransformComponent'
 
 describe('Structure of Array Synchronization', () => {
